third_party/WebKit/LayoutTests/editing/execCommand/hilitecolor.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        '<div contenteditable>foo ^bar| baz</div>',
        'hilitecolor red',
        [
            '<div contenteditable>',
                'foo <span style="background-color: red;">^bar|</span> baz',
            '</div>',
        ].join('')),
        'set background color for "bar"');
</script>
